@@ -20,7 +20,7 @@
         "skype": "jagepard.ru"
     },
     "require": {
-        "php": ">=7.2"
+        "php": ">=7.4"
     },
     "require-dev"      : {
         "phpunit/phpunit": "^7"
